server/pyproject.toml
[project]
name = "naukrigraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.13",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
